+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - CreditNote.payment method
 - IssuingBillingInvoice resource
 - IssuingBillingTransaction resource
+- IssuingDesign resource
+- IssuingEmbossingKit resource
+- IssuingEmbossingRequest resource
+- IssuingEmbossingRequest.Log sub-resource
 - IssuingRule missing parameters
 ### Changed
 - IssuingBin resource to IssuingProduct